New preview command, added missing start & end intervals
- Preview command creates a preview audio file with audible bleeps
placed at beat locations. This will give the user an idea of where
scene changes will occur in the music video for a given song.
- Added missing start and end intervals to the beat intervals provided
by essentia. essentia provides the beat intervals between actual beat
ticks, but not between the first beat tick/beginning of the song, or
the last beat tick/end of the song.
- Added new audio module folder